pyproject.toml
[tool.poetry]
name = "estatapi"
version = "0.0.0"
description = ""
authors = ["savioursho <savioursho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pydantic = "^2.6.4"
pandas = "^2.2.2"
lxml = "^5.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = {extras = ["jupyter"], version = "^23.7.0"}
jupyter = "^1.0.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=estatapi --cov-branch -s -vv"
testpaths = [
    "tests",
]